[ALPS04768139]  Thermal: Fix Coverity issue
[Detail]
Fix Unintentional integer overflow.

MTK-Commit-Id: fd36a57ca07a99bf4c64f7af629d62bafe455409

CR-Id: ALPS04768139
Feature: Thermal Management
(cherry picked from commit 4c4fc58428d1ba56e4612a83b5d89627cabbee61)

+++ drivers/misc/mediatek/thermal/mt6885/src/mtk_lvts_tc.c
@@ -901,8 +901,8 @@ static unsigned int lvts_temp_to_raw(int temp, enum lvts_sensor_enum ts_name)
 	 */
 	unsigned int msr_raw = 0;
 
-	msr_raw = ((long long int)((g_golden_temp * 500 + LVTS_COEFF_B_X_1000
-			- temp)) << 14)/(-1 * LVTS_COEFF_A_X_1000);
+	msr_raw = ((long long int)(((long long int)g_golden_temp * 500 +
+		LVTS_COEFF_B_X_1000 - temp)) << 14)/(-1 * LVTS_COEFF_A_X_1000);
 
 	lvts_printk("%s msr_raw = 0x%x,temp=%d\n", __func__, msr_raw, temp);
